refactor : stations exception 경로 이동
