+++ src/main/java/subway/line/LineService.java
@@ -3,12 +3,11 @@ package subway.line;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import subway.station.Station;
+import subway.station.StationNotFoundException;
 import subway.station.StationRepository;
 import subway.station.StationResponse;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 
 @Service
