+++ src/main/java/subway/station/StationService.java
@@ -2,7 +2,6 @@ package subway.station;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.line.StationNotFoundException;
 
 import java.util.List;
 import java.util.stream.Collectors;
